Improve Molecule testing
With the way the test was previously written, if we added support for
Amazon Linux and did not modify the tests then they would pass.  This
is because Amazon Linux would not match _any_ of the if clauses that
check host.system_info.distribution, and hence nothing would be
tested.

With the way the tests are written now they will fail if run against a
new distribution such as Amazon Linux.  This will prompt us to modify
the tests so that distribution is tested appropriately.